IV-48 refactor: notification을 feature에서 core로 이동
