+++ core/notification/proguard-rules.pro
