+++ core/notification/src/main/AndroidManifest.xml
