+++ core/notification/src/main/java/com/dev/innerview/core/notification/NotificationHelper.kt
@@ -1,4 +1,4 @@
-package com.dev.innerview.feature.notification
+package com.dev.innerview.core.notification
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
