+++ core/notification/src/main/res/drawable-anydpi/ic_stat_name.xml
