+++ core/notification/src/main/res/drawable-hdpi/ic_stat_name.png
