+++ core/notification/src/main/res/drawable-mdpi/ic_stat_name.png
