+++ core/notification/src/main/res/drawable-xhdpi/ic_stat_name.png
