+++ core/notification/src/main/res/drawable-xxhdpi/ic_stat_name.png
