+++ feature/home/build.gradle.kts
@@ -7,5 +7,4 @@ android {
 }
 
 dependencies {
-    implementation(project(":feature:notification"))
 }
