--- feature/notification/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-    id("innerview.android.feature")
-}
-
-android {
-    namespace = "com.dev.innerview.feature.notification"
-}
-
-dependencies {
-}
