+++ settings.gradle.kts
@@ -28,6 +28,7 @@ include(":feature:home")
 include(":feature:peek")
 include(":feature:edit")
 include(":feature:profile")
+include(":feature:record")
 
 include(":core:designsystem")
 include(":core:navigation")
@@ -36,5 +37,4 @@ include(":core:model")
 include(":core:data")
 include(":core:data-api")
 include(":core:domain")
-include(":feature:record")
-include(":feature:notification")
+include(":core:notification")
